pyproject.toml
[project]
name = "hotel-booking-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Vlad Zavorotnyy", email = "vlad.zavorotnyy@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest (>=8.3.5,<9.0.0)",
    "pre-commit (>=4.1.0,<5.0.0)",
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]

[tool.poetry]
packages = [{ include = "core", from = "src" }]

[tool.poetry.scripts]
manage = "src.manage:main"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.1.0"
pytest-django = "^4.10.0"
mypy = "^1.15.0"
django-stubs = "^5.1.3"
djangorestframework-stubs = "^3.15.3"
ruff = "^0.11.0"
isort = "^6.0.1"
pytest-cov = "^6.0.0"
black = "^25.1.0"
drf-yasg = "^1.21.7"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
strict = true
check_untyped_defs = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.isort]
profile = "black"
known_first_party = ["hotels", "bookings", "config"]
line_length = 120
skip = ["migrations", "__pycache__"]

[tool.ruff]
line-length = 120
exclude = ["migrations", "__pycache__"]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["W293", "W391", "E501"]
